eslint wip
