+++ .eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+    "extends": "airbnb",
+    "installedESLint": true,
+    "parser": "babel-eslint",
+    "plugins": [
+        "react"
+    ]
+};
