+++ package.json
@@ -57,10 +57,16 @@
   },
   "homepage": "https://github.com/krujos/willitconnect#readme",
   "devDependencies": {
+    "babel-eslint": "^6.1.2",
     "babel-plugin-transform-runtime": "^6.9.0",
     "babel-preset-es2015": "^6.9.0",
     "chai": "^3.5.0",
     "enzyme": "^2.4.1",
+    "eslint": "^2.13.1",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.12.0",
+    "eslint-plugin-jsx-a11y": "^1.5.5",
+    "eslint-plugin-react": "^5.2.2",
     "expect": "^1.20.2",
     "fs": "0.0.2",
     "karma": "^1.1.1",
