+++ src/main/script/EntryList.js
