+++ src/main/script/HeaderBar.js
