+++ src/main/script/test/EntryBox.spec.jsx
