+++ src/main/script/test/HeaderBar.spec.jsx
