+++ tests.webpack.js
@@ -2,5 +2,5 @@ const chai = require('chai');
 
 global.assert = chai.assert;
 
-const context = require.context('./src/main/scripts/test', true, /.+\.spec\.jsx?$/);
+const context = require.context('./src/main/script/test', true, /.+\.spec\.jsx?$/);
 context.keys().forEach(context);
